services/room-13-fastapi-coordinator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==41.0.7
aiohttp==3.9.1
httpx==0.25.2
redis==5.0.1
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
sqlalchemy-utils==0.41.1
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
PyJWT==2.10.1
